Completo

+++ backend/controllers/tareasControllers.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Tarea = require("../models/tareasModel");
 
 const getTareas = asyncHandler(async (req, res) => {
-  const tareas = await Tarea.find();
+  const tareas = await Tarea.find({ user: req.user._id });
   return res.status(200).json(tareas);
 });
 
@@ -13,6 +13,7 @@ const createTareas = asyncHandler(async (req, res) => {
   }
   const tarea = await Tarea.create({
     texto: req.body.texto,
+    user: req.user._id,
   });
   res.status(201).json(tarea);
 });
@@ -23,10 +24,19 @@ const updateTareas = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("La Tarea NO existe");
   }
-  const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-  });
-  res.status(200).json(tareaUpdated);
+  if (tarea.user.toString !== req.user._id.toString) {
+    res.status(401);
+    throw new Error("Acceso NO Autorizado");
+  } else {
+    const tareaUpdated = await Tarea.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      {
+        new: true,
+      }
+    );
+    res.status(200).json(tareaUpdated);
+  }
 });
 
 const deleteTareas = asyncHandler(async (req, res) => {
@@ -35,14 +45,15 @@ const deleteTareas = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("La Tarea NO existe");
   }
-
-  tarea.deleteOne()
-  // const tareaDeleted = await Tarea.findByIdAndDelete(req.params.id)
-  res.status(200).json({id: req.params.id})
+  if (tarea.user.toString !== req.user._id.toString) {
+    res.status(401);
+    throw new Error("Acceso NO Autorizado");
+  } else {
+    tarea.deleteOne();
+    // const tareaDeleted = await Tarea.findByIdAndDelete(req.params.id)
+    res.status(200).json({ id: req.params.id });
+  }
 });
-
-
-
 
 module.exports = {
   getTareas,
